+++ src/lib.rs
@@ -80,7 +80,7 @@ pub fn process_file(
     );
     let parser = Parser::new_ext(
         &input,
-        Options::ENABLE_TASKLISTS | Options::ENABLE_STRIKETHROUGH,
+        Options::ENABLE_TASKLISTS | Options::ENABLE_STRIKETHROUGH | Options::ENABLE_TABLES,
     );
     let env = Environment::for_local_directory(&base_dir)?;